requirements.txt
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
idna==2.6
matplotlib==2.1.2
numpy==1.14.1
Pillow==5.0.0
praw==5.3.0
prawcore==0.13.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
update-checker==0.16
urllib3==1.22
wordcloud==1.3.1
